requirements.txt
opencv-python
pytesseract
pillow
flask